idle is no longer a Watcher method since only Playwright uses it, and it doesn't need access to other class variables
