+++ src/resource-archive/index.ts
@@ -1,8 +1,5 @@
-import type { Page } from 'playwright';
 import type { Protocol } from 'playwright-core/types/protocol';
 import { logger } from '../utils/logger';
-
-import { DEFAULT_GLOBAL_RESOURCE_ARCHIVE_TIMEOUT_MS } from '../constants';
 
 export type UrlString = string;
 
@@ -51,38 +48,7 @@ export class Watcher {
 
   async watch() {
     this.client.on('Fetch.requestPaused', this.requestPaused.bind(this));
-
     await this.client.send('Fetch.enable');
-  }
-
-  async idle(page: Page, networkTimeoutMs = DEFAULT_GLOBAL_RESOURCE_ARCHIVE_TIMEOUT_MS) {
-    let globalNetworkTimerId: null | ReturnType<typeof setTimeout> = null;
-    let globalNetworkResolver: null | (() => void) = null;
-    // XXX_jwir3: The way this works is as follows:
-    // There are two promises created here. They wrap two separate timers, and we await on a race of both Promises.
-
-    // The first promise wraps a global timeout, where all requests MUST complete before that timeout has passed.
-    // If the timeout passes, an error is thrown. This promise can only throw errors, it cannot resolve successfully.
-    const globalNetworkTimeout = new Promise<void>((resolve) => {
-      globalNetworkResolver = resolve;
-
-      globalNetworkTimerId = setTimeout(() => {
-        logger.warn(`Global timeout of ${networkTimeoutMs}ms reached`);
-        globalNetworkResolver();
-      }, networkTimeoutMs);
-    });
-
-    // The second promise wraps a network idle timeout. This uses playwright's built-in functionality to detect when the network
-    // is idle.
-    const networkIdlePromise = page.waitForLoadState('networkidle').finally(() => {
-      clearTimeout(globalNetworkTimerId);
-    });
-
-    await Promise.race([globalNetworkTimeout, networkIdlePromise]);
-  }
-
-  setResponse(url: UrlString, response: ArchiveResponse) {
-    this.archive[url] = response;
   }
 
   async clientSend<T extends keyof Protocol.CommandParameters>(
